requirements.txt
pymongo
pandas
dash
jupyter-dash
dash-leaflet
plotly
matplotlib
numpy
python-dotenv